requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
sqlalchemy
sqlalchemy-timescaledb
psycopg2-binary
garminconnect
fit-tool
celery
python-telegram-bot
python-dotenv
alembic
cryptography
pydantic-settings
